chore: extract protocol utils to separate files
